requirements.txt
pytest==8.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
prometheus-api-client==0.5.4
schema==0.7.5